Move tests of pattern in its dedicated file
